third_party/WebKit/LayoutTests/xr/events_session_end.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../xr/resources/xr-device-mocking.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<canvas id="webgl-canvas"></canvas>

<script>

let fakeDevices = fakeXRDevices();
let watcherDone = new Event("watcherdone");

xr_session_promise_test( (session, t) => {
  let eventWatcher = new EventWatcher(t, session, ["end", "watcherdone"]);
  let eventPromise = eventWatcher.wait_for(["end", "watcherdone"]);

  function onSessionEnd(event) {
    t.step( () => {
      assert_equals(event.session, session);
      session.dispatchEvent(watcherDone);
    });
  }
  session.addEventListener("end", onSessionEnd, false);
  session.end();

  return eventPromise;
}, fakeDevices["FakeGooglePixelPhone"], [
      { exclusive: true },
      { outputContext: getOutputContext() }
  ],
"Test end fires when session ends");

</script>
